feat(utils): add splitTagToArray that will transform a tag into an array alpha and num splited
